Remove WASM builds after wasm2es6js processing
This removes the generated WASM files (`.wasm` & `_bg.wasm`) after performing a `wasm2es6js` build.

This prevents Webpack from resolving and trying to bundle these files instead of their equally named JS files, outputted from the transformation. If Webpack attempts to bundle these WASM files, it will fail, as they are not a WASM format it can import, and are imported synchronously, which is not supported at the time of this commit.

This also increases the required version of `wasm-bindgen` to 0.2.0, which brings in new features and fixes.

This also updates and fixes an error in the wasm-bindgen example, where the incorrect options were passed to the loader.